Revert "Plan 9 from Bell Labs Fourth Edition 2005-11-20"
This reverts commit 73e742d79f6b0cfc24f3b01d7ade790955db63c2.

+++ sys/src/cmd/rc/win32.c
@@ -307,7 +307,7 @@ Execute(word *args, word *path)
 int
 Globsize(char *p)
 {
-	int isglob = 0, globlen = NDIR+1;
+	ulong isglob = 0, globlen = NDIR+1;
 	for(;*p;p++){
 		if(*p==GLOB){
 			p++;
